:art: move file mutations to their own component
This untangles the file mutation components from the FileList container, helping us to crawl out of the deep nesting
